// film-network-ui/src/app/services/models/film-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface FilmResponse {
  addedBy?: number;
  archive?: boolean;
  createdAt?: string;
  director?: string;
  filmPoster?: string;
  genre?: 'ACTION' | 'ADVENTURE' | 'ANIMATION' | 'BIOGRAPHY' | 'COMEDY' | 'CRIME' | 'DOCUMENTARY' | 'DRAMA' | 'FAMILY' | 'FANTASY' | 'HISTORY' | 'HORROR' | 'MUSICAL' | 'MYSTERY' | 'ROMANCE' | 'SCI_FI' | 'SPORT' | 'THRILLER' | 'WAR' | 'WESTERN';
  id?: number;
  imdbId?: string;
  lastModifiedAt?: string;
  rating?: number;
  synopsis?: string;
  title?: string;
  year?: string;
}