// film-network-ui/src/app/services/models/page-response-rented-film-response.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { RentedFilmResponse } from '../models/rented-film-response';
export interface PageResponseRentedFilmResponse {
  content?: Array<RentedFilmResponse>;
  first?: boolean;
  last?: boolean;
  pageNumber?: number;
  size?: number;
  totalElements?: number;
  totalPages?: number;
}